pyproject.toml
[project]
name = "agentic-data-analyst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "anthropic>=0.79.0",
    "ipykernel>=7.2.0",
    "langgraph>=1.0.9",
    "logfire>=4.24.0",
    "mcp>=1.26.0",
    "openai>=2.21.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "plotly>=6.5.2",
    "pydantic-ai>=1.59.0",
    "python-dotenv>=1.2.1",
    "python-pptx>=1.0.2",
    "tabulate>=0.9.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::logfire._internal.config.LogfireNotConfiguredWarning",
]
